==> Data Preprocessing/3_Excluded_MCDI_Items.rmd <==

# MCDI 
## Data Exclusion

```{r, exclude_mcdi}
# Remove undesired item types.
exclude_types <- c(
  "how_use_words",
  "complexity",
  "word_endings",
  "combine",
  "word_endings_nouns",
  "word_endings_verbs"
)

mcdi <- raw_mcdi %>% filter(!type %in% exclude_types)

#remove definitions based on the following criteria:
#1)Remove from data if definition is compound word (e.g play dough).
#2)Include only a single item for words with multiple meaning (e.g can(object, can (auxiliary)))

remove_items <-
  c(
    "baa baa",
    "choo choo",
    "quack quack",
    "uh oh",
    "woof woof",
    "yum yum",
    "play dough",
    "chicken (food)",
    "fish (food)",
    "french fries",
    "green beans",
    "ice cream",
    "peanut butter",
    "potato chip",
    "belly button",
    "high chair",
    "living room",
    "rocking chair",
    "washing machine",
    "play pen",
    "lawn mower",
    "gas station",
    "babysitter's name",
    "child's own name",
    "pet's name",
    "give me five!",
    "gonna get you!",
    "go potty",
    "night night",
    "so big!",
    "thank you",
    "this little piggy",
    "turn around",
    "drink (action)",
    "slide (action)",
    "swing (action)",
    "watch (action)",
    "work (action)",
    "all gone",
    "clean (description)",
    "dry (description)",
    "orange (description)",
    "next to",
    "on top of",
    "a lot",
    "can (auxiliary)",
    "don't",
    "water (not beverage)"
  )

mcdi <- mcdi %>% filter(!definition %in% remove_items)

#Replaces specific word meaning (e.g "drink (beverage)" with lemma (drink))
mcdi$definition <-
  recode(
    mcdi$definition,
    "chicken (animal)" = "chicken",
    "fish (animal)" = "fish",
    "toy (object)" = "toy",
    "drink (beverage)" = "drink",
    "orange (food)" = "orange",
    "water (beverage)" = "water",
    "dress (object)" = "dress",
    "buttocks/bottom*" = "bottom",
    "owie/boo boo" = "owie",
    "penis*" = "penis",
    "vagina*" = "vagina",
    "can (object)" = "can",
    "tissue/kleenex" = "tissue",
    "watch (object)" = "watch",
    "slide (object)" = "slide",
    "swing (object)" = "swing" ,
    "church*" = "church",
    "work (place)" = "work",
    "daddy*" = "daddy",
    "grandma*" = "grandma",
    "grandpa*" = "granpa",
    "mommy*" = "mommy",
    "call (on phone)" = "call",
    "clean (action)" = "clean",
    "dry (action)" = "dry",
    "little (description)" = "little",
    "inside/in" = "inside",
    "did/did ya" = "did",
    "gonna/going to" = "gonna",
    "gotta/got to" = "gotta",
    "hafta/have to" = "hafta",
    "lemme/let me" = "lemme",
    "need/need to" = "need",
    "try/try to" = "try",
    "wanna/want to" = "wanna",
    "shh/shush/hush" = "hush",
    "soda/pop" = "soda"
  )


#Save data after excluded items
write.csv(mcdi, file = "clean_mcdi.csv")
```
